requirements.txt
annotated-types==0.7.0
argcomplete==3.5.3
bitarray==3.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
ckzg==2.0.1
click==8.1.8
colorama==0.4.6
cytoolz==1.0.1
distlib==0.3.9
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
filelock==3.17.0
hexbytes==1.3.0
idna==3.10
iniconfig==2.0.0
packaging==24.2
parsimonious==0.10.0
pipx==1.7.1
platformdirs==4.3.6
pluggy==1.5.0
poly_eip712_structs==0.0.1
py_clob_client==0.20.0
py_order_utils==0.3.2
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.4
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
toolz==1.0.0
typing_extensions==4.12.2
urllib3==2.3.0
userpath==1.9.2
virtualenv==20.29.2
